+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-<!--next-version-placeholder-->
+
+## v0.4.0 (06/14/2025)
+
+- Ovelapping indels are handled according to the same strategy as bcftools consensus. After an indel is applied that position is frozen and no other variants can be applied to the same region.
+
+- Test cases for variant application
+
+
 
 ## v0.1.0 (01/05/2025)